Reducing the borders of the player so that he does not touch the background of the game.

+++ src/utils/HelpMethods.java
@@ -3,7 +3,7 @@ package utils;
 import main.Game;
 
 public class HelpMethods {
-    public static boolean canMoveHere(float x, float y, int width, int height, int[][] levelData) {
+    public static boolean canMoveHere(float x, float y, float width, float height, int[][] levelData) {
         if (!isSolid(x, y, levelData))
             if (!isSolid(x + width, y + height, levelData))
                 if (!isSolid(x + width, y, levelData))
